Added cdn caching logic
